refactor ptreferential to externalize the legacy implementation
