+++ source/ptreferential/CMakeLists.txt
@@ -1,4 +1,10 @@
-SET(PTREF_SRC ptreferential.cpp ptreferential_api.cpp where.h reflexion.h ptref_graph.cpp)
+SET(PTREF_SRC
+  ptreferential.cpp
+  ptreferential_legacy.cpp
+  ptreferential_api.cpp
+  where.h
+  reflexion.h
+  ptref_graph.cpp)
 add_library(ptreferential ${PTREF_SRC})
 
 add_subdirectory(tests)
