+++ source/ptreferential/ptreferential.cpp
@@ -1,0 +1,67 @@
+/* Copyright © 2001-2014, Canal TP and/or its affiliates. All rights reserved.
+  
+This file is part of Navitia,
+    the software to build cool stuff with public transport.
+ 
+Hope you'll enjoy and contribute to this project,
+    powered by Canal TP (www.canaltp.fr).
+Help us simplify mobility and open public transport:
+    a non ending quest to the responsive locomotion way of traveling!
+  
+LICENCE: This program is free software; you can redistribute it and/or modify
+it under the terms of the GNU Affero General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+   
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+GNU Affero General Public License for more details.
+   
+You should have received a copy of the GNU Affero General Public License
+along with this program. If not, see <http://www.gnu.org/licenses/>.
+  
+Stay tuned using
+twitter @navitia 
+IRC #navitia on freenode
+https://groups.google.com/d/forum/navitia
+www.navitia.io
+*/
+
+#include "ptreferential.h"
+#include "ptreferential_legacy.h"
+
+namespace navitia{
+namespace ptref{
+
+type::Indexes make_query(const type::Type_e requested_type,
+                         const std::string& request,
+                         const std::vector<std::string>& forbidden_uris,
+                         const type::OdtLevel_e odt_level,
+                         const boost::optional<boost::posix_time::ptime>& since,
+                         const boost::optional<boost::posix_time::ptime>& until,
+                         const type::Data& data) {
+    return make_query_legacy(requested_type, request, forbidden_uris, odt_level, since, until, data);
+}
+
+type::Indexes make_query(const type::Type_e requested_type,
+                         const std::string& request,
+                         const std::vector<std::string>& forbidden_uris,
+                         const type::Data &data) {
+    return make_query(requested_type,
+                      request,
+                      forbidden_uris,
+                      type::OdtLevel_e::all,
+                      boost::none,
+                      boost::none,
+                      data);
+}
+
+type::Indexes make_query(const type::Type_e requested_type,
+                         const std::string& request,
+                         const type::Data &data) {
+    const std::vector<std::string> forbidden_uris;
+    return make_query(requested_type, request, forbidden_uris, data);
+}
+
+}}// namespace navitia::ptref
